converted character class to int

+++ app/src/main/java/com/example/wallproject/Model/Dungeon/Characters/Character.kt
@@ -12,8 +12,8 @@ abstract class Character{
     //transient is exlucing values of being serialized by the GSON class for JSON
     @Transient private val defenseThreshold = 0.8
     @Transient private val criticalMultiplier = 1.5
-    @Transient private val hundred = 100.0
-    @Transient private val zero = 0.0
+    @Transient private val hundred = 100
+    @Transient private val zero = 0
 
     fun isDead(): Boolean {
         return (health <= zero)
@@ -24,8 +24,8 @@ abstract class Character{
         var trueAttack = calculateAttackTrue(character)
 
         if(isCritical(trueAttack.luck)){
-            trueAttack.attackDmg = trueAttack.attackDmg * criticalMultiplier
-            trueAttack.abilityPower = trueAttack.abilityPower * criticalMultiplier
+            trueAttack.attackDmg = (trueAttack.attackDmg * criticalMultiplier).toInt()
+            trueAttack.abilityPower = (trueAttack.abilityPower * criticalMultiplier).toInt()
         }
 
         character.health = character.health - (trueAttack.attackDmg + trueAttack.abilityPower)
@@ -36,9 +36,9 @@ abstract class Character{
 
     private fun calculateDefensePercentage(character: Character) : Defense {
 
-        var resultArmor = ruleOfThree(character.health, hundred, character.defense.armor) / hundred
+        var resultArmor = ruleOfThree(character.health, hundred.toDouble(), character.defense.armor.toDouble()) / hundred
 
-        var resultMagicResist = ruleOfThree(character.health, hundred, character.defense.magicResist) / hundred
+        var resultMagicResist = ruleOfThree(character.health, hundred.toDouble(), character.defense.magicResist.toDouble()) / hundred
 
         if(resultArmor > defenseThreshold)
             resultArmor = defenseThreshold
@@ -46,12 +46,12 @@ abstract class Character{
         if (resultMagicResist > defenseThreshold)
             resultMagicResist = defenseThreshold
 
-        return Defense(resultArmor, resultMagicResist)
+        return Defense(resultArmor.toInt(), resultMagicResist.toInt())
 
     }
 
     private fun calculateLuckPercentage() : Double {
-        return ruleOfThree(this.health, hundred, this.attack.luck)
+        return ruleOfThree(this.health, hundred.toDouble(), this.attack.luck.toDouble())
     }
 
     private fun calculateAttackTrue(character: Character) : Attack {
@@ -64,16 +64,16 @@ abstract class Character{
 
         resultAttack.abilityPower = this.attack.abilityPower * resultDefense.magicResist
 
-        resultAttack.luck = calculateLuckPercentage()
+        resultAttack.luck = calculateLuckPercentage().toInt()
 
         return resultAttack
 
     }
 
     //function calculates chance for a critical strike
-    private fun isCritical (luck : Double) : Boolean {
+    private fun isCritical (luck : Int) : Boolean {
 
-        var result = Random.nextDouble(hundred + 1)
+        var result = Random.nextInt(hundred + 1)
 
         return (result < luck)
